TMVA Harmonise generation of GUI histograms
The histogram generation for binary/multiclass classification and
regression do not follow the same code paths. This patch makes the
generation more similar for the binary and multiclass classification
cases. (Regression is left unchanged for now).

CrossEvaluation requires the histograms to be generated with going
through the Factory method TestAllMethods, where in the multiclass
case _some_ of the histograms where previously generated.